Meet with Ajay and Jade and fix sql issue

+++ backend/models/office_hours/ticket_statistics.py
@@ -14,9 +14,9 @@ class OfficeHoursTicketStatistics(BaseModel):
     of the `OfficeHoursTicket` database in the PostgreSQL database.
     """
 
-    total_tickets: int
-    total_tickets_weekly: int
-    average_wait_time: int
-    average_duration: int
-    total_conceptual: int
-    total_assignment: int
+    total_tickets: float
+    total_tickets_weekly: float
+    average_wait_time: float
+    average_duration: float
+    total_conceptual: float
+    total_assignment: float
